+++ src/device/wocurtain.ts
@@ -7,6 +7,42 @@ import { Buffer } from 'buffer';
 import { SwitchbotDevice } from '../switchbot.js';
 
 export class WoCurtain extends SwitchbotDevice {
+  static parseServiceData(buf, onlog) {
+    if (buf.length !== 5 && buf.length !== 6) {
+      if (onlog && typeof onlog === 'function') {
+        onlog(
+          `[parseServiceDataForWoCurtain] Buffer length ${buf.length} !== 5 or 6!`,
+        );
+      }
+      return null;
+    }
+    const byte1 = buf.readUInt8(1);
+    const byte2 = buf.readUInt8(2);
+    const byte3 = buf.readUInt8(3);
+    const byte4 = buf.readUInt8(4);
+
+    const calibration = byte1 & 0b01000000 ? true : false; // Whether the calibration is compconsted
+    const battery = byte2 & 0b01111111; // %
+    const inMotion = byte3 & 0b10000000 ? true : false;
+    const currPosition = byte3 & 0b01111111; // current positon %
+    const lightLevel = (byte4 >> 4) & 0b00001111; // light sensor level (1-10)
+    const deviceChain = byte4 & 0b00000111;
+    const model = buf.slice(0, 1).toString('utf8');
+
+    const data = {
+      model: model,
+      modelName: 'WoCurtain',
+      calibration: calibration,
+      battery: battery,
+      inMotion: inMotion,
+      position: currPosition,
+      lightLevel: lightLevel,
+      deviceChain: deviceChain,
+    };
+
+    return data;
+  }
+
   /* ------------------------------------------------------------------
    * open()
    * - Open the curtain
@@ -70,7 +106,7 @@ export class WoCurtain extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target position percentage is incorrect: ' +
-              typeof percent,
+            typeof percent,
           ),
         );
       });
