+++ src/device/woplugmini.ts
@@ -6,6 +6,84 @@ import { SwitchbotDevice } from '../switchbot.js';
  * @see https://github.com/OpenWonderLabs/SwitchBotAPI-BLE/blob/latest/devicetypes/plugmini.md
  */
 export class WoPlugMini extends SwitchbotDevice {
+  static parseServiceData_US(manufacturerData, onlog) {
+    if (manufacturerData.length !== 14) {
+      if (onlog && typeof onlog === 'function') {
+        onlog(
+          `[parseServiceDataForWoPlugMiniUS] Buffer length ${manufacturerData.length} should be 14`,
+        );
+      }
+      return null;
+    }
+    const byte9 = manufacturerData.readUInt8(9); // byte9:  plug mini state; 0x00=off, 0x80=on
+    const byte10 = manufacturerData.readUInt8(10); // byte10: bit0: 0=no delay,1=delay, bit1:0=no timer, 1=timer; bit2:0=no sync time, 1=sync'ed time
+    const byte11 = manufacturerData.readUInt8(11); // byte11: wifi rssi
+    const byte12 = manufacturerData.readUInt8(12); // byte12: bit7: overload?
+    const byte13 = manufacturerData.readUInt8(13); // byte12[bit0~6] + byte13: current power value
+
+    const state = byte9 === 0x00 ? 'off' : byte9 === 0x80 ? 'on' : null;
+    const delay = !!(byte10 & 0b00000001);
+    const timer = !!(byte10 & 0b00000010);
+    const syncUtcTime = !!(byte10 & 0b00000100);
+    const wifiRssi = byte11;
+    const overload = !!(byte12 & 0b10000000);
+    const currentPower = (((byte12 & 0b01111111) << 8) + byte13) / 10; // in watt
+    // TODO: voltage ???
+
+    const data = {
+      model: 'g',
+      modelName: 'WoPlugMini',
+      state: state,
+      delay: delay,
+      timer: timer,
+      syncUtcTime: syncUtcTime,
+      wifiRssi: wifiRssi,
+      overload: overload,
+      currentPower: currentPower,
+    };
+
+    return data;
+  }
+
+  static parseServiceData_JP(manufacturerData, onlog) {
+    if (manufacturerData.length !== 14) {
+      if (onlog && typeof onlog === 'function') {
+        onlog(
+          `[parseServiceDataForWoPlugMiniJP] Buffer length ${manufacturerData.length} should be 14`,
+        );
+      }
+      return null;
+    }
+    const byte9 = manufacturerData.readUInt8(9); // byte9:  plug mini state; 0x00=off, 0x80=on
+    const byte10 = manufacturerData.readUInt8(10); // byte10: bit0: 0=no delay,1=delay, bit1:0=no timer, 1=timer; bit2:0=no sync time, 1=sync'ed time
+    const byte11 = manufacturerData.readUInt8(11); // byte11: wifi rssi
+    const byte12 = manufacturerData.readUInt8(12); // byte12: bit7: overload?
+    const byte13 = manufacturerData.readUInt8(13); // byte12[bit0~6] + byte13: current power value
+
+    const state = byte9 === 0x00 ? 'off' : byte9 === 0x80 ? 'on' : null;
+    const delay = !!(byte10 & 0b00000001);
+    const timer = !!(byte10 & 0b00000010);
+    const syncUtcTime = !!(byte10 & 0b00000100);
+    const wifiRssi = byte11;
+    const overload = !!(byte12 & 0b10000000);
+    const currentPower = (((byte12 & 0b01111111) << 8) + byte13) / 10; // in watt
+    // TODO: voltage ???
+
+    const data = {
+      model: 'j',
+      modelName: 'WoPlugMini',
+      state: state,
+      delay: delay,
+      timer: timer,
+      syncUtcTime: syncUtcTime,
+      wifiRssi: wifiRssi,
+      overload: overload,
+      currentPower: currentPower,
+    };
+
+    return data;
+  }
+
   /**
    * @returns {Promise<boolean>} resolves with a boolean that tells whether the plug in ON(true) or OFF(false)
    */
