+++ src/device/wostrip.ts
@@ -6,6 +6,55 @@ import { SwitchbotDevice } from '../switchbot.js';
  * @see https://github.com/OpenWonderLabs/SwitchBotAPI-BLE/blob/latest/devicetypes/colorbulb.md
  */
 export class WoStrip extends SwitchbotDevice {
+  static parseServiceData(buf, onlog) {
+    if (buf.length !== 18) {
+      if (onlog && typeof onlog === 'function') {
+        onlog(
+          `[parseServiceDataForWoStrip] Buffer length ${buf.length} !== 18!`,
+        );
+      }
+      return null;
+    }
+
+    //const byte1 = buf.readUInt8(1);//power and light status
+    //const byte2 = buf.readUInt8(2);//bulb brightness
+    const byte3 = buf.readUInt8(3);//bulb R
+    const byte4 = buf.readUInt8(4);//bulb G
+    const byte5 = buf.readUInt8(5);//bulb B
+    const byte7 = buf.readUInt8(7);
+    const byte8 = buf.readUInt8(8);
+    const byte9 = buf.readUInt8(9);
+    const byte10 = buf.readUInt8(10);
+
+    const state = byte7 & 0b10000000 ? true : false;
+    const brightness = byte7 & 0b01111111;
+    const red = byte3;
+    const green = byte4;
+    const blue = byte5;
+    const delay = byte8 & 0b10000000;
+    const preset = byte8 & 0b00001000;
+    const color_mode = byte8 & 0b00000111;
+    const speed = byte9 & 0b01111111;
+    const loop_index = byte10 & 0b11111110;
+
+    const data = {
+      model: 'r',
+      modelName: 'WoStrip',
+      state: state,
+      brightness: brightness,
+      red: red,
+      green: green,
+      blue: blue,
+      delay: delay,
+      preset: preset,
+      color_mode: color_mode,
+      speed: speed,
+      loop_index: loop_index,
+    };
+
+    return data;
+  }
+
   /**
    * @returns {Promise<boolean>} resolves with a boolean that tells whether the plug in ON(true) or OFF(false)
    */
@@ -44,7 +93,7 @@ export class WoStrip extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target brightness percentage is incorrect: ' +
-              typeof brightness,
+            typeof brightness,
           ),
         );
       });
@@ -66,7 +115,7 @@ export class WoStrip extends SwitchbotDevice {
         reject(
           new Error(
             'Strip Light Doesn\'t Support Color temperature: ' +
-              typeof color_temperature,
+            typeof color_temperature,
           ),
         );
       });
@@ -82,7 +131,7 @@ export class WoStrip extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target brightness percentage is incorrect: ' +
-              typeof brightness,
+            typeof brightness,
           ),
         );
       });
@@ -92,7 +141,7 @@ export class WoStrip extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target red is incorrect: ' +
-              typeof red,
+            typeof red,
           ),
         );
       });
@@ -102,7 +151,7 @@ export class WoStrip extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target green is incorrect: ' +
-              typeof green,
+            typeof green,
           ),
         );
       });
@@ -112,7 +161,7 @@ export class WoStrip extends SwitchbotDevice {
         reject(
           new Error(
             'The type of target blue is incorrect: ' +
-              typeof blue,
+            typeof blue,
           ),
         );
       });
@@ -165,7 +214,7 @@ export class WoStrip extends SwitchbotDevice {
             reject(
               new Error(
                 'Expecting a 2-byte response, got instead: 0x' +
-                  res_buf.toString('hex'),
+                res_buf.toString('hex'),
               ),
             );
           }
