+++ src/switchbot.ts
@@ -13,17 +13,18 @@ import { WoHumi } from './device/wohumi.js';
 import { WoPlugMini } from './device/woplugmini.js';
 import { WoBulb } from './device/wobulb.js';
 import { WoStrip } from './device/wostrip.js';
+
 type params = {
-    duration?: number,
-    model?: string,
-    id?: string,
-    quick?: false,
-    noble?: any,
+  duration?: number,
+  model?: string,
+  id?: string,
+  quick?: false,
+  noble?: any,
 }
 
 type peripherals = {
-    addr?: string,
-    id?: string,
+  addr?: string,
+  id?: string,
 }
 
 export class SwitchBot {
@@ -167,7 +168,7 @@ export class SwitchBot {
       this._init()
         .then(() => {
           const peripherals: peripherals = {};
-          let timer: NodeJS.Timeout = setTimeout(() => {}, 0);
+          let timer: NodeJS.Timeout = setTimeout(() => { }, 0);
           const finishDiscovery = () => {
             if (timer) {
               clearTimeout(timer);
@@ -300,7 +301,7 @@ export class SwitchBot {
             device = new WoPlugMini(peripheral, this.noble);
             break;
           case 'o':
-          //device = new SwitchbotDeviceWoSmartLock(peripheral, this.noble);
+            //device = new SwitchbotDeviceWoSmartLock(peripheral, this.noble);
             break;
           case 'i':
             device = new WoSensorTH(peripheral, this.noble);
@@ -453,7 +454,7 @@ export class SwitchBot {
             if (SwitchBot.filterAdvertising(ad, p.id, p.model)) {
               if (
                 this.onadvertisement &&
-                                typeof this.onadvertisement === 'function'
+                typeof this.onadvertisement === 'function'
               ) {
                 this.onadvertisement(ad);
               }
